# perf(q): switch for each loops to for loops in extractor-t