--- a/plugins/q/src/data/extractor-t.js
+++ b/plugins/q/src/data/extractor-t.js
@@ -140,8 +140,8 @@ function doIt({
   ret,
   sourceKey
 }) {
-  propsArr.forEach((prop) => {
-    const pCfg = props[prop];
+  for (let i = 0; i < propsArr.length; i++) {
+    const pCfg = props[propsArr[i]];
     const arr = pCfg.fields || [pCfg];
     let coll;
     let collStr;
@@ -149,29 +149,30 @@ function doIt({
       coll = [];
       collStr = [];
     }
-    arr.forEach((p) => {
+
+    for (let j = 0; j < arr.length; j++) {
       let fn;
       let str;
       let value;
       let nodes;
       let cells;
       let label;
-      if (p.type === 'primitive') {
-        value = p.value;
-        label = String(p.value);
+      if (arr[j].type === 'primitive') {
+        value = arr[j].value;
+        label = String(arr[j].value);
       } else {
-        if (typeof p.value === 'function') {
-          fn = v => p.value(v, item);
+        if (typeof arr[j].value === 'function') {
+          fn = v => arr[j].value(v, item);
         }
-        if (typeof p.label === 'function') {
-          str = v => p.label(v, item);
+        if (typeof arr[j].label === 'function') {
+          str = v => arr[j].label(v, item);
         }
-        if (p.accessor) {
-          nodes = p.accessor(item);
+        if (arr[j].accessor) {
+          nodes = arr[j].accessor(item);
           if (Array.isArray(nodes)) { // propably descendants
-            cells = nodes.map(p.valueAccessor);
-            if (p.attrAccessor) {
-              cells = cells.map(p.attrAccessor);
+            cells = nodes.map(arr[j].valueAccessor);
+            if (arr[j].attrAccessor) {
+              cells = cells.map(arr[j].attrAccessor);
             }
             if (fn) {
               value = cells.map(fn);
@@ -181,10 +182,10 @@ function doIt({
               label = cells.map(str);
               str = null;
             }
-            value = p.reduce ? p.reduce(value) : value;
-            label = p.reduceLabel ? p.reduceLabel(label, value) : String(value);
+            value = arr[j].reduce ? arr[j].reduce(value) : value;
+            label = arr[j].reduceLabel ? arr[j].reduceLabel(label, value) : String(value);
           } else {
-            value = p.attrAccessor ? p.attrAccessor(p.valueAccessor(nodes)) : p.valueAccessor(nodes);
+            value = arr[j].attrAccessor ? arr[j].attrAccessor(arr[j].valueAccessor(nodes)) : arr[j].valueAccessor(nodes);
             label = value;
           }
         } else {
@@ -198,22 +199,22 @@ function doIt({
         collStr.push(str && label != null ? str(label) : (label != null ? label : String(v)));
       } else {
         const v = fn ? fn(value) : value;
-        ret[prop] = {
+        ret[propsArr[i]] = {
           value: v,
           label: str ? str(label) : (label != null ? label : String(v))
         };
-        if (p.field) {
-          ret[prop].source = { field: p.field.key(), key: sourceKey };
+        if (arr[j].field) {
+          ret[propsArr[i]].source = { field: arr[j].field.key(), key: sourceKey };
         }
       }
-    });
+    }
     if (coll) {
-      ret[prop] = {
+      ret[propsArr[i]] = {
         value: typeof pCfg.value === 'function' ? pCfg.value(coll, item) : coll,
         label: typeof pCfg.label === 'function' ? pCfg.label(collStr, item) : collStr
       };
     }
-  });
+  }
 }
 
 const getHierarchy = (cube, cache, config) => {
@@ -284,19 +285,19 @@ const attachPropsAccessors = ({
   itemDepthObject,
   f
 }) => {
-  propsArr.forEach((prop) => {
-    const pCfg = props[prop];
+  for (let i = 0; i < propsArr.length; i++) {
+    const pCfg = props[propsArr[i]];
     const arr = pCfg.fields ? pCfg.fields : [pCfg];
-    arr.forEach((p) => {
-      if (p.field !== f) {
-        const depthObject = getFieldDepth(p.field, { cube, cache });
+    for (let j = 0; j < arr.length; j++) {
+      if (arr[j].field !== f) {
+        const depthObject = getFieldDepth(arr[j].field, { cube, cache });
         const accessors = getFieldAccessor(itemDepthObject, depthObject);
-        p.accessor = accessors.nodeFn; // nodes accessor
-        p.valueAccessor = accessors.valueFn; // cell accessor
-        p.attrAccessor = accessors.attrFn; // attr cell accessor
+        arr[j].accessor = accessors.nodeFn; // nodes accessor
+        arr[j].valueAccessor = accessors.valueFn; // cell accessor
+        arr[j].attrAccessor = accessors.attrFn; // attr cell accessor
       }
-    });
-  });
+    }
+  }
 };
 
 export function augment(config = {}, dataset, cache, util) {
@@ -350,12 +351,12 @@ export function augment(config = {}, dataset, cache, util) {
   const replicaDescendants = replica.descendants();
   const descendants = h.descendants();
 
-  descendants.forEach((node, idx) => {
-    const propsArr = propDefs[node.depth].propsArr;
-    const props = propDefs[node.depth].props;
-    const main = propDefs[node.depth].main;
+  for (let i = 0; i < descendants.length; i++) {
+    const propsArr = propDefs[descendants[i].depth].propsArr;
+    const props = propDefs[descendants[i].depth].props;
+    const main = propDefs[descendants[i].depth].main;
 
-    const item = replicaDescendants[idx];
+    const item = replicaDescendants[i];
     const itemData = item.data; // main.valueAccessor(currentOriginal);
 
     const ret = datumExtract(main, itemData, { key: sourceKey });
@@ -368,25 +369,25 @@ export function augment(config = {}, dataset, cache, util) {
       sourceKey,
       isTree: true
     });
-    node.data = ret;
-  });
+    descendants[i].data = ret;
+  }
   return h;
 }
 
 export function extract(config, dataset, cache, util) {
   const cfgs = Array.isArray(config) ? config : [config];
   let dataItems = [];
-  cfgs.forEach((cfg) => {
-    if (typeof cfg.field !== 'undefined') {
+  for (let g = 0; g < cfgs.length; g++) {
+    if (typeof cfgs[g].field !== 'undefined') {
       const cube = dataset.raw();
       const sourceKey = dataset.key();
       const h = getHierarchy(cube, cache, config);
       if (!h) {
-        return;
+        continue;
       }
 
-      const f = typeof cfg.field === 'object' ? cfg.field : dataset.field(cfg.field);
-      const { props, main } = util.normalizeConfig(cfg, dataset);
+      const f = typeof cfgs[g].field === 'object' ? cfgs[g].field : dataset.field(cfgs[g].field);
+      const { props, main } = util.normalizeConfig(cfgs[g], dataset);
       const propsArr = Object.keys(props);
 
       const itemDepthObject = getFieldDepth(f, { cube, cache });
@@ -401,8 +402,8 @@ export function extract(config, dataset, cache, util) {
         f
       });
 
-      const track = !!cfg.trackBy;
-      const trackType = typeof cfg.trackBy;
+      const track = !!cfgs[g].trackBy;
+      const trackType = typeof cfgs[g].trackBy;
       const tracker = {};
       const trackedItems = [];
 
@@ -428,7 +429,7 @@ export function extract(config, dataset, cache, util) {
         // items with the same trackBy value are placed in an array and reduced later
         if (track) {
           util.track({
-            cfg,
+            cfg: cfgs[g],
             itemData,
             obj: ret,
             target: trackedItems,
@@ -449,6 +450,6 @@ export function extract(config, dataset, cache, util) {
         dataItems.push(...mapped);
       }
     }
-  });
+  }
   return dataItems;
 }