**Checklist**

- [ ] tests added
- [x] commits conform to the [commit guidelines](./CONTRIBUTING.md#commit)
- [ ] documentation updated

- By switching for each loops to regular for loops inside extractor-t.js the following results was yielded when running performance benchmarks. This was tested on around 500 field values.

Extraction on qDimensionInfo/0

for each loops
~2700 ops/s

for loops
~ 4851 ops/s

Extraction on qDimensionInfo/1

for each loops
~1697 ops/s

for loops
~ 3000 ops/s


- In order to check out the metrics, please check out the branch located on this link and follow the instructions in README.md https://github.com/nguyenlam1/picasso.js/tree/perf-extract-t/plugins/q/src/data